+++ api/Filer.Application/Services/PostService.cs
@@ -1,4 +1,5 @@
 using Filer.Application.Interfaces;
+using Filer.DataAccess;
 using Filer.DataAccess.Interfaces;
 using Filer.Domain.Domain;
 namespace Filer.Application.Services;
@@ -12,12 +13,20 @@ public class PostService : IPostService
     }
     public async Task Create(Post post)
     {
+        var existedUser = await postRepository.Get(post.Creator!.Id);
+        if(existedUser.Id == 0){
+            return;
+        }
         await postRepository.Create(post);
         await postRepository.Save();
     }
 
     public async Task Delete(int id)
     {
+        var post = await postRepository.Get(id);
+        if(post.Id == 0){
+            return;
+        }
         await postRepository.Delete(id);
         await postRepository.Save();
     }
@@ -34,6 +43,14 @@ public class PostService : IPostService
 
     public async Task Update(Post post)
     {
+        var existedPost = await postRepository.Get(post.Id);
+        if(existedPost.Id == 0){
+            return;
+        }
+        var existedUser = await postRepository.Get(post.Creator!.Id);
+        if(existedUser.Id == 0){
+            return;
+        }
         await postRepository.Update(post);
         await postRepository.Save();
     }
